+++ Project/src/main/java/project/controller/RegistrationController.java
@@ -4,25 +4,19 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.Errors;
-import org.springframework.validation.FieldError;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import project.model.EntityDtoConverter;
+import project.model.EntityDtoConverterOlolo;
 import project.dto.UserDTO;
 import project.entity.RoleType;
 import project.entity.User;
-import project.exceptions.DuplicatedEmailException;
 import project.service.UserService;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.Map;
-import java.util.Optional;
 
 @Controller
 public class RegistrationController {
@@ -30,8 +24,6 @@ public class RegistrationController {
     private ControllerUtils controllerUtils;
     @Autowired
     private UserService userService;
-    @Autowired
-    private EntityDtoConverter entityDtoConverter;
 
     @GetMapping(value = {"/registration"})
     public String showRegistrationForm(Model model, @RequestParam(required = false, defaultValue = "false") Boolean success,
