+++ Project/src/main/java/project/controller/UsersAdministrationController.java
@@ -5,14 +5,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import project.entity.User;
-import project.model.EntityDtoConverter;
+import project.model.EntityDtoConverterOlolo;
 import project.dto.UserDTO;
 import project.entity.RoleType;
 import project.service.UserService;
@@ -26,8 +25,6 @@ import java.util.List;
 public class UsersAdministrationController {
     @Autowired
     private UserService userService;
-    @Autowired
-    private EntityDtoConverter entityDtoConverter;
 
     @RequestMapping()
     public String usersPage(Model model) {
