+++ Project/src/main/java/project/model/MoneyFormatConverter.java
@@ -1,26 +1,11 @@
 package project.model;
 
-import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.stereotype.Component;
 import project.controller.GlobalConstants;
 
-import java.util.Locale;
+public interface MoneyFormatConverter {
+    long getMoneyValue(String money);
 
-@Component
-public class MoneyFormatConverter {
-    public long getMoneyValue(String money) {
-        try {
-            return (long) (Double.parseDouble(money) * GlobalConstants.MONEY_TO_VALUE_COEFFICIENT);
-        }catch (NumberFormatException e){
-            return 0;
-        }
-    }
+    String getStringMoneyFromMoneyValue(Long moneyValue);
 
-    public String getStringMoneyFromMoneyValue(Long moneyValue) {
-        return String.format("%.2f",(double)moneyValue/100);
-    }
-
-    public long getMoneyValue(double money) {
-        return (long) (money * GlobalConstants.MONEY_TO_VALUE_COEFFICIENT);
-    }
+    long getMoneyValue(double money);
 }
