+++ Project/src/main/java/project/repository/AccountRepository.java
@@ -21,7 +21,6 @@ public interface AccountRepository extends JpaRepository<Account, Long> {
 
     Optional<Account> findByAccountName(String accountName);
 
-//    @Modifying
-//    @Query("update Account a set a.money = a.money+?2 where a.id = ?1")
-//    void updateMoneyById(Long id, long money);
+    @Query("select a from Account a left join CreditCard c on c.account.id=a.id where a.user.id = :userId and c.account.id is null")
+    List<Account> findFreeUserAccountsByUserId(Long userId);
 }
