+++ Project/src/main/java/project/service/AccountService.java
@@ -18,7 +18,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-interface AccountService {
+public interface AccountService {
 
     Account save(Account account);
